requirements.txt
Flask
opencv
numpy
werkzeug
keras
tensorflow
pillow
os